+++ packages/vscode-messenger-common/src/index.ts
@@ -5,4 +5,5 @@
  ******************************************************************************/
 
 export * from './messages';
+export * from './util';
 export * from './cancellation';